refactor(modes): prepare codebase for various modes
