+++ src/index.constants.ts
@@ -1,6 +1,7 @@
 import type { StampOptions } from "@/index.types";
 
 const DEFAULT_STAMP_OPTIONS: StampOptions = {
+  mode: "comment",
   targetSelector: "body",
 } as const;
