+++ src/index.spec.ts
@@ -1,6 +1,7 @@
-import { Node, Window } from "happy-dom";
+import { Window } from "happy-dom";
 import { describe } from "vitest";
 
+import * as CommentMode from "@/modes/comment/comment";
 import { stampInHtml } from "@/index";
 
 const window = new Window({ url: "https://localhost:8080" });
@@ -12,6 +13,7 @@ describe("Dev Stamp Index", () => {
       globalThis.window = window as unknown as typeof globalThis.window;
       globalThis.document = document as unknown as Document;
       window.document.body.innerHTML = "<body><h1>Title</h1><p>Text</p></body>";
+      vi.spyOn(CommentMode, "stampCommentInHtml").mockImplementation(vi.fn());
     });
 
     it("should log an error when not in a browser environment because window is undefined.", () => {
@@ -37,22 +39,11 @@ describe("Dev Stamp Index", () => {
       expect(consoleErrorSpy).toHaveBeenCalledWith("Target element not found: #nonexistent");
     });
 
-    it("should append a comment node to the target element when called.", () => {
+    it("should stamp comment in html when called.", () => {
       const message = "Hello Dark Jess' 🪄";
       stampInHtml(message);
-      const targetElement = window.document.querySelector("body");
-      const appendedNode = targetElement?.lastChild;
 
-      expect(appendedNode?.nodeType).toBe(Node.COMMENT_NODE);
-    });
-
-    it("should append a comment node with the message specified in first argument to the target element when called.", () => {
-      const message = "Hello Dark Jess' 🪄";
-      stampInHtml(message);
-      const targetElement = window.document.querySelector("body");
-      const appendedNode = targetElement?.lastChild;
-
-      expect(appendedNode?.nodeValue).toBe(message);
+      expect(CommentMode.stampCommentInHtml).toHaveBeenCalledWith(message, window.document.body);
     });
   });
 });
