+++ src/index.ts
@@ -1,5 +1,6 @@
 import type { StampOptions } from "@/index.types";
 import { DEFAULT_STAMP_OPTIONS } from "@/index.constants";
+import { stampCommentInHtml } from "@/modes/comment/comment";
 import { getStampOptions } from "@/utils/utils";
 
 function stampInHtml(message: string, options: Partial<StampOptions> = DEFAULT_STAMP_OPTIONS): void {
@@ -16,8 +17,7 @@ function stampInHtml(message: string, options: Partial<StampOptions> = DEFAULT_S
 
     return;
   }
-  const commentNode = window.document.createComment(message);
-  targetElement.appendChild(commentNode);
+  stampCommentInHtml(message, targetElement);
 }
 
 export {
