+++ src/modes/comment/comment.ts
@@ -1,0 +1,6 @@
+function stampCommentInHtml(message: string, targetElement: Element): void {
+  const commentNode = window.document.createComment(message);
+  targetElement.appendChild(commentNode);
+}
+
+export { stampCommentInHtml };
